test(csa-session): keep daemon completion Available regression local
Move the Available-session daemon completion regression into the daemon completion module so the final branch diff does not grow the oversized session result test module.
