+++ crates/cli-sub-agent/src/session_cmds_result_tests.rs
@@ -368,77 +368,6 @@ fn handle_session_result_uses_daemon_completion_outcome() {
     assert_eq!(persisted.phase, csa_session::SessionPhase::Retired);
 }
 
-#[cfg(unix)]
-#[test]
-fn handle_session_result_preserves_available_session_on_daemon_completion() -> anyhow::Result<()> {
-    let tmp = tempdir()?;
-    let _env_lock = TEST_ENV_LOCK.blocking_lock();
-    let state_home = tmp.path().join("xdg-state");
-    std::fs::create_dir_all(&state_home)?;
-    let _home_guard = EnvVarGuard::set("HOME", tmp.path());
-    let _state_guard = EnvVarGuard::set("XDG_STATE_HOME", &state_home);
-    let project = tmp.path();
-
-    let session = create_session(
-        project,
-        Some("result-available-completion-packet"),
-        None,
-        Some("codex"),
-    )?;
-    let session_id = session.meta_session_id;
-    let session_dir = get_session_dir(project, &session_id)?;
-
-    let mut persisted = csa_session::load_session(project, &session_id)?;
-    persisted.phase = csa_session::SessionPhase::Available;
-    csa_session::save_session(&persisted)?;
-
-    let now = chrono::Utc::now();
-    let existing = SessionResult {
-        status: "success".to_string(),
-        exit_code: 0,
-        summary: "existing compact result".to_string(),
-        tool: "codex".to_string(),
-        original_tool: None,
-        fallback_tool: None,
-        fallback_reason: None,
-        started_at: now,
-        completed_at: now,
-        events_count: 0,
-        artifacts: Vec::new(),
-        peak_memory_mb: None,
-        fallback_chain: None,
-        gate_timeout: false,
-        warnings: Vec::new(),
-        raw_process_exit_code: None,
-        uncommitted_changes: None,
-        manager_fields: Default::default(),
-    };
-    csa_session::save_result(project, &session_id, &existing)?;
-    let result_path = session_dir.join(csa_session::result::RESULT_FILE_NAME);
-    let result_before = std::fs::read_to_string(&result_path)?;
-
-    std::fs::write(
-        session_dir.join("daemon-completion.toml"),
-        "exit_code = 17\nstatus = \"failure\"\n",
-    )?;
-
-    handle_session_result(
-        session_id.clone(),
-        false,
-        Some(project.to_string_lossy().into_owned()),
-        StructuredOutputOpts::default(),
-    )?;
-
-    let persisted = csa_session::load_session(project, &session_id)?;
-    assert_eq!(persisted.phase, csa_session::SessionPhase::Available);
-    assert_eq!(std::fs::read_to_string(&result_path)?, result_before);
-
-    let result = load_result(project, &session_id)?.expect("existing result should remain");
-    assert_eq!(result.status, "success");
-    assert_eq!(result.exit_code, 0);
-    Ok(())
-}
-
 #[cfg(target_os = "linux")]
 #[test]
 fn handle_session_result_ignores_completion_packet_while_daemon_alive() -> anyhow::Result<()> {
